Module dvm_net::prelude

The DVM-NET Prelude

The purpose of this module is to give imports-set for many common cases by adding a glob import to the top:

use dvm_net::prelude::*;

Also contains reshared TryInto and TryFrom std traits.

Re-exports

pub use crate::serve::*;

pub use crate::endpoint::*;

pub use crate::transport::*;

Traits

TryFrom
Simple and safe type conversions that may fail in a controlled way under some circumstances. It is the reciprocal of TryInto.

TryInto
An attempted conversion that consumes self, which may or may not be expensive.